crowd-counting-experiments/experiments/cifar10-cnn-classifier/requirements.txt
flask>=2.0.0
torch>=1.10.0
torchvision>=0.11.0
Pillow>=8.0.0
numpy>=1.21.0